refactor(core): extract app factory — eliminate bottom-of-file router imports (#124)
main.py previously imported all 25+ routers at the bottom of the file to
work around circular imports (routers importing from main.py).  This made
the file both an app definition and a shared-state container.

New src/core/app_factory.py owns:
  - _init_sentry()            — optional Sentry SDK initialisation
  - lifespan()                — asyncpg + aioredis pool lifecycle
  - _register_exception_handlers() — HTTP / unhandled / rate-limit handlers
  - _register_middleware()    — CorrelationId + CORS
  - _register_routers()       — all 25+ router imports (lazy, inside function
                                body) + include_router calls + dev_router gate
  - create_app()              — assembles and returns the FastAPI instance

main.py is now 3 meaningful lines:
  from src.core.app_factory import create_app
  app = create_app()

All existing test files import `from main import app` — that still works
since main.py exports app at module level.  No test changes required.
